Separate all the platform independent code
